pyproject.toml
[tool.poetry]
name = "ssa"
version = "0.1.0"
description = ""
authors = ["PAL <info@predictive-analytics-lab.com>"]

[tool.poetry.dependencies]
python = "^3.9"
neoconfigen = {git = "https://github.com/predictive-analytics-lab/neoconfigen.git", branch = "master"}
hydra-submitit-launcher = {version = "^1.1.1", allow-prereleases = true}
numpy = "^1.21.1"
pandas = "^1.3.1"
palbolts = {git = "https://github.com/predictive-analytics-lab/palbolts.git", branch = "main" }
palkit = "^0.3.8"
pytorch-lightning = "^1.3.8"
sklearn = "^0.0"
tqdm = "*"
wandb = "^0.11.0"
gpytorch = "^1.5.0"

[tool.poetry.dev-dependencies]
black = "*"
mypy = "*"
pydocstyle = "*"
pylint = "*"
pytest = "*"
pytest-cov = "*"
pre-commit = "*"
python-type-stubs = {git = "https://github.com/predictive-analytics-lab/python-type-stubs.git", branch = "main" }


[tool.poetry.scripts]
gogo = "ssa.main:launcher"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 100
target-version = ['py38']
skip-string-normalization = true
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
known_third_party = ["absl", "cv2", "google", "kit", "matplotlib", "numba", "numpy", "pandas", "pytest", "pytorch_lightning", "sdc", "sklearn", "tensorflow", "torch", "tqdm", "transformers", "transforms3d", "wandb", "ysdc_dataset_api"]
known_future_library = []
extra_standard_library = ["dataclasses", "__future__", "typing_extensions"]
line_length = 88
profile = "black"
force_sort_within_sections = "True"
classes = ["MISSING"]

[tool.pyright]
include = ["ssa"]
exclude = ["**/node_modules", "**/__pycache__"]
typeCheckingMode = "basic"
pythonVersion = "3.7"
reportUnusedImport = "error"
reportDuplicateImport  = "error"
reportIncompatibleVariableOverride = "error"
reportOverlappingOverload = "error"
reportUntypedNamedTuple = "error"
reportConstantRedefinition = "warning"
reportMissingImports = "error"
reportMissingTypeStubs = "warning"
strictListInference = true
strictSetInference = true
strictParameterNoneValue = true

[tool.mypy]
python_version = "3.8"
no_implicit_optional = true
allow_redefinition = true
strict_equality = true
check_untyped_defs = true
warn_unreachable = true
disallow_any_unimported = false
disallow_untyped_defs = true
disallow_incomplete_defs = true
show_error_codes = true
warn_unused_ignores = false

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "pytest.*",
    "sklearn.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = []
follow_imports = "skip"
follow_imports_for_stubs = true
